+++ packages/dart/CHANGELOG.md
@@ -7,6 +7,7 @@
 - Add `deleteFile` method
 - Add `createFile` method
 - Add `deleteAllFileFromDirectory` method
+- Add `moveFile` method
 
 ## 1.1.7 (2025-11-03)